# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: TransitIntel API
  version: 0.2.0
  description: TransitIntel — Public transport planning intelligence platform for Provincia di Ancona
servers:
  - url: /api
    description: Base API path
security: []   # Public API — no authentication required
...
paths:
...
  /cron/weather:
    post:
      operationId: cronWeather
      tags: [cron, weather]
      summary: Fetch weather snapshots from OpenWeatherMap for province locations
      parameters:
        - name: x-cron-secret
          in: header
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  inserted: { type: integer }
                  errors: { type: integer }
...
  /weather/history:
    get:
      operationId: getWeatherHistory
      tags: [weather]
      summary: Weather history (configurable lookback)
      parameters:
        - name: hours
          in: query
          schema: { type: integer, default: 24 }
          description: Number of hours to look back
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeatherSnapshot"
  /weather/correlation:
    get:
      operationId: getWeatherCorrelation
      tags: [weather]
      summary: Correlation between weather conditions and traffic congestion
      parameters:
        - name: hours
          in: query
          schema: { type: integer, default: 168 }
          description: Number of hours to look back for correlation analysis
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    weather_main: { type: string }
                    sample_count: { type: integer }
                    avg_congestion: { type: number }
                    avg_speed_kmh: { type: number }
                    avg_temp: { type: number }
                    avg_humidity: { type: number }
                    avg_wind_speed: { type: number }

  # ── Admin ─────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    WeatherSnapshot:
      type: object
      properties:
        id: { type: string, format: uuid }
        lat: { type: number }
        lng: { type: number }
        locationName: { type: string }
        temp: { type: number, description: Temperature in Celsius }
        feelsLike: { type: number }
        humidity: { type: integer }
        windSpeed: { type: number, description: Wind speed m/s }
        weatherMain: { type: string, description: "Main weather group (Clear, Rain, Snow, etc.)" }
        weatherDescription: { type: string }
        weatherIcon: { type: string }
        rain1h: { type: [number, "null"], description: "Rain volume last 1h mm" }
        snow1h: { type: [number, "null"], description: "Snow volume last 1h mm" }
        visibility: { type: [integer, "null"], description: "Visibility in meters" }
        capturedAt: { type: string, format: date-time }